requirements.txt
asgiref==3.3.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
Django==3.1.3
django-cors-headers==3.5.0
django-email-utils==1.0.0
django-encrypted-filefield==0.3.0
django-hosts==4.0
django-nested-admin==3.3.2
djangorestframework==3.12.2
emailutils==0.0.1
ffmpeg==1.4
idna==2.10
pycparser==2.20
python-ffmpeg-video-streaming==0.1.14
python-magic==0.4.18
python-monkey-business==1.0.0
pytz==2020.4
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
django-environ==0.4.5
psycopg2-binary==2.8.6
pandas==1.1.4
xlrd==1.2.0
google-api-python-client==1.12.8
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
oauth2client==4.1.3
uWSGI>=2.0.19.1,<2.1.0
flake8>=3.9.2,<3.10.0
django-extensions>=3.1.3,<3.2.0
idpay>=0.0.1,<0.1.0
Faker>=8.4.0,<8.5.0